requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.1
azure-identity==1.14.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
django==4.2.4
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.4
msal==1.23.0
msal-extensions==1.0.0
multidict==6.0.4
numpy==1.25.2
openai==0.27.8
openpyxl==3.1.2
pandas==2.0.3
pip==23.2.1
portalocker==2.7.0
pycparser==2.21
pyjwt==2.8.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
requests==2.31.0
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.4


wheel==0.38.4
whitenoise==6.5.0

yarl==1.9.2

